chore: support generic rules to have routers drain events
