+++ router/batchrouter/types.go
@@ -6,34 +6,11 @@ import (
 	backendconfig "github.com/rudderlabs/rudder-server/backend-config"
 	"github.com/rudderlabs/rudder-server/jobsdb"
 	router_utils "github.com/rudderlabs/rudder-server/router/utils"
-	"github.com/rudderlabs/rudder-server/utils/misc"
 )
 
 type Connection struct {
 	Source      backendconfig.SourceT
 	Destination backendconfig.DestinationT
-}
-
-type JobParameters struct {
-	SourceID                string `json:"source_id"`
-	DestinationID           string `json:"destination_id"`
-	ReceivedAt              string `json:"received_at"`
-	TransformAt             string `json:"transform_at"`
-	SourceTaskRunID         string `json:"source_task_run_id"`
-	SourceJobID             string `json:"source_job_id"`
-	SourceJobRunID          string `json:"source_job_run_id"`
-	SourceDefinitionID      string `json:"source_definition_id"`
-	DestinationDefinitionID string `json:"destination_definition_id"`
-	SourceCategory          string `json:"source_category"`
-	EventName               string `json:"event_name"`
-	EventType               string `json:"event_type"`
-	MessageID               string `json:"message_id"`
-}
-
-// ParseReceivedAtTime parses the [ReceivedAt] field and returns the parsed time or a zero value time if parsing fails
-func (jp *JobParameters) ParseReceivedAtTime() time.Time {
-	receivedAt, _ := time.Parse(misc.RFC3339Milli, jp.ReceivedAt)
-	return receivedAt
 }
 
 type DestinationJobs struct {
